Rework rangesAreDisjoint
I think that the previous version of rangesAreDisjoint was not quite
correct, or maybe I don't understand what that function was trying to
accomplish (it's hard to figure out one way or another).

I reimplemented rangesAreDisjoint so that it works the way I think it
should work, and wrote tests for it.