Refactor Range to be a sum: PosRange | LineRange
This changes the core representation of a Range from a pair of positions
to *either* a pair of positions (PosRange) *or* a pair of line numbers
(LineRange). This is a fairly fundamental change, which is why many
files were touched, but luckily we have types.

In particular, I'd like to call attention to:

- I made Range concrete (so that you can pattern match on it) but made
  PosRange and LineRange abstract, so that you have to go through smart
  constructors to build them.

- I couldn't think of a good name for the Range constructors, so I just
  called them PR and LR. I'm open to alternative names here.