Update README.src.md, and update CSS
In addition to updating the README, I updated the docs page's CSS to
take advantage of the new inline / block CSS classes we're now applying.
I did this so that people could more easily discover this feature
(because it does require some extra CSS changes and thus more work on
their end).

If you'd prefer to instead keep the examples as is and just mention the
CSS bit in a footnote somewhere, that's fine by me too.